rename edit helper, quick fix for use_new
