+++ packages/core/src/edit_helper.ts
@@ -34,4 +34,10 @@ export class EditHelper {
     const range: IRange = {start: pos, end: pos};
     return {[filename]: [{range, newText: text}]};
   }
+
+  public static replaceRange(file: IFile, start: Position, end: Position, text: string): IEdit {
+    const filename = file.getFilename();
+    const range: IRange = {start, end};
+    return {[filename]: [{range, newText: text}]};
+  }
 }
