+++ packages/core/src/rules/empty_statement.ts
@@ -3,7 +3,7 @@ import {Empty} from "../abap/2_statements/statements/_statement";
 import {ABAPRule} from "./_abap_rule";
 import {ABAPFile} from "../files";
 import {BasicRuleConfig} from "./_basic_rule_config";
-import {EditHelper} from "../edit";
+import {EditHelper} from "../edit_helper";
 import {Position} from "../position";
 
 export class EmptyStatementConf extends BasicRuleConfig {
